Split tests into submodules
