+++ test/acceptance.spec.js
@@ -1,6 +1,5 @@
 import { expect } from 'chai'
-import { returnSuggestions, getCursorPosition, getReferenceString } from '../src/index.js'
-import { stringMatch } from '../src/string_matching.js'
+import { returnSuggestions } from '../src/index.js'
 
 describe('returnSugestions (Acceptance Tests)', () => {
     it('should return empty list if file is empty', () => {
@@ -163,68 +162,5 @@ describe('returnSugestions (Acceptance Tests)', () => {
         let user = new UsɅ
         `);
         expect(result).to.include("User")
-    });
-});
-
-describe('getCursorPosition', () => {
-    it('should return cursor position given cursor signified by caret symbol', () => {
-        const result = getCursorPosition(`const foo = cɅ`);
-        expect(result).to.equal(13)
-    });
-
-    it('should return -1 if caret symbol not found', () => {
-        const result = getCursorPosition(`const foo = c`);
-        expect(result).to.equal(-1)
-    });
-});
-
-describe('getReferenceString', () => {
-    it('should return empty string if whitespace found before cursor pos', () => {
-        const file = "const foo Ʌ"
-        const referenceString = getReferenceString(file, 10);
-        expect(referenceString).to.equal('')
-    });
-
-    it('should return correct string before cursor if no whitespace', () => {
-        const file = "const foo = barɅ"
-        const referenceString = getReferenceString(file, 15);
-        expect(referenceString).to.equal('bar')
-    });
-});
-
-describe('stringMatching', () => {
-    it('should return empty list if the universe array is empty', () => {
-        const words = []
-        const searchStr = 'abc';
-        const suggestions = stringMatch(words, searchStr)
-        expect(suggestions).to.have.lengthOf(0)
-    });
-
-    it('should return empty list if the search string is empty', () => {
-        const words = ['abc', 'foo', 'bar', 'foobarbuzz', 'zas']
-        const searchStr = '';
-        const suggestions = stringMatch(words, searchStr)
-        expect(suggestions).to.have.lengthOf(0)
-    });
-
-    it('should return exact matches when threshold is set to 0', () => {
-        const words = ['abc', 'foo', 'bar', 'foobarbuzz', 'zas']
-        const searchStr = 'foo';
-        const suggestions = stringMatch(words, searchStr, 0)
-        expect(suggestions).to.deep.equal(['foo', 'foobarbuzz'])
-    });
-
-    it('should return reasonable matches when threshold is set to 0.5', () => {
-        const words = ['foibarbuzz', 'abc', 'foo', 'bar', 'foobarbuzz', 'zas']
-        const searchStr = 'foo';
-        const suggestions = stringMatch(words, searchStr, 0.5)
-        expect(suggestions).to.deep.equal(['foo', 'foobarbuzz', 'foibarbuzz'])
-    });
-
-    it('should return reasonable matches ordered by scope when threshold is set to 0.5 and orderedByScope is True', () => {
-        const words = ['foibarbuzz', 'abc', 'foo', 'bar', 'foobarbuzz', 'zas']
-        const searchStr = 'foo';
-        const suggestions = stringMatch(words, searchStr, 0.5, true)
-        expect(suggestions).to.deep.equal(['foibarbuzz', 'foo', 'foobarbuzz'])
     });
 });
